backend/requirements.txt
alembic==1.10.4
anyio==3.6.2
asyncpg==0.27.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
ecdsa==0.18.0
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
loguru==0.7.0
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.10
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
win32-setctime==1.1.0
